+++ problem/allocation.go
@@ -1,6 +1,7 @@
 package problem
 
 import (
+	"github.com/roidaradal/fn/dict"
 	"github.com/roidaradal/fn/ds"
 	"github.com/roidaradal/fn/list"
 	"github.com/roidaradal/opt/data"
@@ -16,6 +17,10 @@ func NewAllocation(variant string, n int) *discrete.Problem {
 		return resourceAllocation(name)
 	case "scene":
 		return sceneAllocation(name)
+	case "fair_item":
+		return fairItemAllocation(name)
+	case "house":
+		return houseAllocation(name)
 	default:
 		return nil
 	}
@@ -95,5 +100,106 @@ func sceneAllocation(name string) *discrete.Problem {
 
 	//p.SolutionCoreFn = fn.CoreSortedPartition(domain, cfg.Scenes)
 	p.SolutionStringFn = fn.StringPartition(domain, cfg.Scenes)
+	return p
+}
+
+// Common steps for creating Item Allocation problem
+func newItemAllocationProblem(name string) (*discrete.Problem, *data.ItemAllocation) {
+	cfg := data.NewItemAllocation(name)
+	if cfg == nil {
+		return nil, nil
+	}
+
+	p := discrete.NewProblem(name)
+	p.Type = discrete.Assignment
+	p.Goal = discrete.Minimize
+
+	p.Variables = discrete.Variables(cfg.Items)
+	domain := discrete.Domain(cfg.Persons)
+	p.AddVariableDomains(domain)
+
+	return p, cfg
+}
+
+// Fair Item Allocation
+func fairItemAllocation(name string) *discrete.Problem {
+	p, cfg := newItemAllocationProblem(name)
+	if p == nil || cfg == nil {
+		return nil
+	}
+	domain := p.UniformDomain()
+
+	p.AddUniversalConstraint(func(solution *discrete.Solution) bool {
+		partitions := fn.AsPartition(solution, domain)
+		nonEmptyPartitions := list.Filter(partitions, func(partition []discrete.Value) bool {
+			return len(partition) > 0
+		})
+		// If there are less items than persons, minimum number of person who received an item should be numItems
+		// If there are less persons than items, minimum number of person who received an item should be numPersons
+		// This ensures that one or few people are not hoarding the items and are being distributed evenly
+		minCount := min(len(cfg.Items), len(cfg.Persons))
+		return len(nonEmptyPartitions) >= minCount
+	})
+
+	p.ObjectiveFn = func(solution *discrete.Solution) discrete.Score {
+		var envy discrete.Score = 0
+		partitions := fn.AsPartition(solution, domain)
+		for x1, p1 := range cfg.Persons {
+			value1 := list.Sum(list.Map(partitions[x1], func(item discrete.Variable) float64 {
+				return cfg.Value[p1][item]
+			}))
+			for x2 := range cfg.Persons {
+				if x1 == x2 {
+					continue
+				}
+				value2 := list.Sum(list.Map(partitions[x2], func(item discrete.Variable) float64 {
+					return cfg.Value[p1][item]
+				}))
+				if value2 > value1 {
+					envy += value2 - value1
+				}
+			}
+		}
+		return envy
+	}
+
+	p.SolutionStringFn = fn.StringPartition(domain, cfg.Items)
+	return p
+}
+
+// House Allocation
+func houseAllocation(name string) *discrete.Problem {
+	p, cfg := newItemAllocationProblem(name)
+	if p == nil || cfg == nil {
+		return nil
+	}
+
+	// Ensure each house is assigned to a different person
+	p.AddUniversalConstraint(fn.ConstraintAllUnique)
+
+	p.ObjectiveFn = func(solution *discrete.Solution) discrete.Score {
+		// Solution.Map contains House => Person, so we swap the dictionary
+		// to form a lookup with Person => House
+		houseOf := dict.Swap(solution.Map)
+		// Sum up envy among pairs of people:
+		// For pair A, B: if A's assigned house has less value to person A
+		// than the the house assigned to B, then A is envious of B's house
+		var envy discrete.Score = 0
+		for x1, p1 := range cfg.Persons {
+			value1 := cfg.Value[p1][houseOf[x1]]
+			for x2 := range cfg.Persons {
+				if x1 == x2 {
+					continue
+				}
+				value2 := cfg.Value[p1][houseOf[x2]]
+				if value2 > value1 {
+					envy += value2 - value1
+				}
+			}
+		}
+		return envy
+	}
+
+	p.SolutionStringFn = fn.StringMap(p, cfg.Items, cfg.Persons)
 	return p
 }